requirements.txt
asgiref==3.4.1
astroid==2.6.6
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.6
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2==2.9.1
PyJWT==2.1.0
pylint==2.9.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.3.0
wrapt==1.12.1
